Fix mt5_account_profiles.login integer overflow for large broker logins
Some brokers issue demo account logins above the 32-bit signed integer
range (e.g. 5054067375 on the new internal_demo_10k account), which caused
a Postgres DataError on the very first real account_registry insert.
Changes the column from Integer to BigInteger in both the ORM model and the
not-yet-shared migration 0030 (safe to edit in place -- unpushed branch,
zero rows existed in the table).

+++ backend/alembic/versions/0030_adaptive_trade_manager_v2.py
@@ -22,7 +22,8 @@ def upgrade() -> None:
     op.create_table(
         "mt5_account_profiles",
         sa.Column("fingerprint_hash", sa.String(64), primary_key=True),
-        sa.Column("login", sa.Integer(), nullable=False),
+        # BigInteger: some brokers issue demo logins above the 32-bit signed range.
+        sa.Column("login", sa.BigInteger(), nullable=False),
         sa.Column("server", sa.String(128), nullable=False),
         sa.Column("company", sa.String(128), nullable=True),
         sa.Column("currency", sa.String(8), nullable=True),
